requirements.txt
Django==1.8.3
djangorestframework==3.5.3
httplib2==0.9.2
oauth==1.0.1
oauth2==1.5.211
oauth2client==1.4.12
oauthlib==2.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.2.12
requests==2.11.1
requests-oauth2==0.2.0
requests-oauthlib==0.5.0
rsa==3.4.2
six==1.10.0
